Authorize in hero controller
[#89854348]

+++ app/controllers/heroes_controller.rb
@@ -5,18 +5,23 @@ class HeroesController < ApplicationController
   before_action :authenticate_user!, except: [:show, :index]
 
   def index
+    authorize hero
   end
 
   def new
+    authorize hero
   end
 
   def edit
+    authorize hero
   end
 
   def show
+    authorize hero
   end
 
   def create
+    authorize hero
     hero.campaign = campaign
     hero.user = current_user
     hero.save
@@ -24,11 +29,13 @@ class HeroesController < ApplicationController
   end
 
   def update
+    authorize hero
     hero.update(hero_params)
     redirect_to campaign_hero_path(campaign, hero)
   end
 
   def destroy
+    authorize hero
     hero.destroy
     redirect_to campaign_heroes_path(campaign)
   end
